Interactive modal
I have added an interactive modal to fill the form of each task. Futhermore, I have also placed a dark background to give the app some volume, although I might change it for the blur effect. On top of this, I keep getting crazy with the height bug on browsers...